refactor: 暂时不暴露 GLOBAL 变量，修改 cr 建议
