+++ jest.setup.js
@@ -1,3 +1,3 @@
-const { setGlobal } = require('./src');
+const { setGlobal } = require('./src/core/global');
 
 setGlobal({ locale: 'zh-CN' });
