+++ src/types/locale.ts
@@ -1,14 +1,14 @@
 export type Locale = {
   locale: string;
 
-  // 通用
+  // 1. 通用
   general: {
     increase: string;
     decrease: string;
     root: string;
   };
 
-  // 按照图表组件
+  // 2. 按照图表组件
   /** 中心文本 */
   statistic: {
     total: string;
@@ -22,9 +22,9 @@ export type Locale = {
   slider?: Record<string, string>;
   scrollbar?: Record<string, string>;
 
-  // 按照图表类型
+  // 3. 按照图表类型
   waterfall: {
-    // 总计或累计值
+    /** 总计或累计值 */
     total: string;
   };
 };
